// src/runtime/server/types/storeApiTypes.d.ts
export type components = {
  schemas: Schemas;
};
export type Schemas = {
// ...
  Cart: {
    /** An affiliate tracking code */
    affiliateCode?: string | null;
    /** @enum {string} */
    apiAlias: 'cart';
    /** A campaign tracking code */
    campaignCode?: string | null;
    /** A comment that can be added to the cart. */
    customerComment?: string | null;
    deliveries?: components['schemas']['CartDelivery'][];
    /** A list of all cart errors, such as insufficient stocks, invalid addresses or vouchers. */
    errors?:
      | components['schemas']['CartError'][]
      | {
          [key: string]: {
            code: number;
            key: string;
            level: number;
            message: string;
            messageKey: string;
          };
        };
    /** All items within the cart */
    lineItems?: components['schemas']['LineItem'][];
    modified?: boolean;
    /** Name of the cart - for example `guest-cart` */
    name?: string;
    price: components['schemas']['CalculatedPrice'];
    /** Context token identifying the cart and the user session */
    token?: string;
    /** A list of all payment transactions associated with the current cart. */
    transactions?: {
      amount?: components['schemas']['CalculatedPrice'];
      paymentMethodId?: string;
    }[];
  };
// ...
};